home page update
home page created with buttons to redirect to upload and retrieve page
