+++ frontend/templates/face.html
@@ -4,8 +4,8 @@
     <title></title>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="{{ url_for('static', filename='fstyles.css') }}">
-    <title>Flask Client Camera Web App</title>
+    <link rel="stylesheet" href="{{ url_for('static', filename='face_styles.css') }}">
+    <title>Face scan</title>
 
     <style>
     #video {
@@ -36,6 +36,6 @@
     </div>
     <h1 id="userName">Unknown</h1>
 
-    <script src="{{ url_for('static', filename='fscript.js') }}"></script>
+    <script src="{{ url_for('static', filename='face_script.js') }}"></script>
   </body>
 </html>
