+++ frontend/templates/upload.html
@@ -1,0 +1,61 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="{{ url_for('static', filename='upload_styles.css') }}">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css">
+    <title>Snap Bank Upload</title>
+</head>
+<body>
+    <form action="/upload_images" method="POST" enctype="multipart/form-data">
+        <div class="container py-3">
+            <div class="panel">
+                <div class="row justify-content-center">
+                    <div class="col-lg-7">
+                        <h1 class="title text-center">
+                            Welcome to Snap bank
+                        </h1>
+                        
+                        <div class="text-right">
+                            <p>upload your <b>Images</b> here.</p>
+                        </div>
+                        <div class="wrapper">
+                            <button type="submit" id="uploadId" class="file-upload"> Upload &nbsp;&nbsp;<i class="fa fa-arrow-up"></i></button>
+                            <div class="clearAllSelection">
+                                <span id="clearAll">Clear All </span>
+                                <i class="fa fa-times clearAllITag"></i>
+
+                            </div>
+                        </div>
+                        
+                    </div>
+                </div>
+                <br>
+                <br>
+                <div class="custom-file-image">
+                    <input type="file" id="fileImage" name="fileImage"
+                            class="custom-file-image-input"
+                            accept="image/*"
+                            multiple="multiple">
+                    <div class="custom-file-preview">
+                        <label class="custom-file-preview-add" for="fileImage">
+                            <i class="fas fa-plus"></i>
+                        </label>
+                    </div>
+                </div>
+                <p class="text-danger error-product_image"></p>
+            </div>
+        </div>
+    </form>
+    {% with messages = get_flashed_messages() %}
+    {% if messages %}
+        {% for message in messages %}
+        <script>alert(`{{message}}`)</script>
+        {% endfor %}
+    {% endif %}
+    {% endwith %}
+    <script src="https://code.jquery.com/jquery-3.7.1.min.js" integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo=" crossorigin="anonymous"></script>
+    <script src="{{ url_for('static', filename='upload_script.js') }}"></script>
+</body>
+</html>
